test(edda-cli): move process-spawning e2e tests to integration tests
Move the process-spawning end-to-end tests from src/cmd_claim.rs,
src/cmd_verify.rs, and src/cmd_ask.rs into dedicated integration
tests under crates/edda-cli/tests/ using env!("CARGO_BIN_EXE_edda").

Previously, running cargo test -p edda --bin edda only compiled the
binary test harness (deps/edda-<hash>.exe) without compiling
target/debug/edda.exe. In a warm lane, cmd_claim.rs would spawn a
stale target/debug/edda.exe from a prior build or fail when absent.
By moving these tests to tests/claim_e2e_contract.rs,
tests/verify_contract.rs, and tests/ask_compat_contract.rs:
- Cargo natively guarantees edda.exe is compiled fresh before running.
- cargo test -p edda --bin edda runs pure unit tests and never depends on
  target/debug/edda.exe.
- cmd_claim.rs line count is reduced from 2540 to 1685, ratcheting down
  the ceiling in scripts/file-length-ceilings.txt.

Issue: #789
