+++ scripts/file-length-ceilings.txt
@@ -13,7 +13,7 @@ crates/edda-bridge-claude/src/dispatch/tests.rs 3098
 crates/edda-bridge-claude/src/peers/tests.rs 3588
 crates/edda-bridge-claude/src/signals.rs 2252
 crates/edda-cli/src/cmd_bridge.rs 3059
-crates/edda-cli/src/cmd_claim.rs 2539
+crates/edda-cli/src/cmd_claim.rs 1685
 crates/edda-cli/src/cmd_dispatch.rs 1996
 crates/edda-cli/src/cmd_draft.rs 1705
 crates/edda-cli/src/cmd_gc.rs 1117
